Transfer function unit tests into TypeScript. https://online.sbis.ru/opendoc.html?guid=9233ac85-d560-4be5-acb3-72fbe908cabc
