+++ TypesUnit/_function/delay.test.ts
@@ -1,0 +1,9 @@
+import delay from 'Types/_function/delay';
+
+describe('Types/_formatter/delay', () => {
+    it('should call method', (done) => {
+        delay(() => {
+            done();
+        });
+    });
+});
