--- TypesUnit/_function/throttle.test.js
@@ -1,20 +1,0 @@
-define([
-   'Types/function'
-], function (
-   functionUtil
-) {
-   describe('Types/_formatter/throttle', () => {
-      it('should call method only one time', (done) => {
-         let value = 1;
-         const decorator = functionUtil.throttle(() => {
-            value += 1;
-         }, 0);
-         decorator();
-         decorator();
-         setTimeout(() => {
-            assert.equal(value, 2);
-            done();
-         }, 150);
-      });
-   });
-});
